Fix review comments from Claudio II.
